testutils: cleaned up testutils
